+++ src/main/java/ua/kpi/nc/service/InterviewService.java
@@ -19,7 +19,7 @@ public interface InterviewService {
 
 	List<Interview> getByApplicationForm(ApplicationForm applicationForm);
 
-	Interview getByApplicationFormAndInterviewerRoleId(ApplicationForm applicationForm, int interviewerRoleId );
+	Interview getByApplicationFormAndInterviewerRoleId(ApplicationForm applicationForm, Long interviewerRoleId );
 
 	Long insertInterview(Interview interview, ApplicationForm applicationForm, User interviewer, Role role);
